Store map values and DP entries as short, add casts for std::min type safety
- Change FastVersionedMap::keyValue[] and operator[] to use short
- Shrink dp[][][] from int to short to cut memory in half
- Wrap std::min operands in matching casts (static_cast<int> or static_cast<short>) to resolve template deduction errors
- Silence narrowing-conversion warnings when assigning back into short arrays
- Tweak Phase 3 ASCII illustrations’ spacing for consistent alignment

+++ include/FastVersionedMap.h
@@ -40,7 +40,7 @@ private:
     int keyVersion[MAX_KEYS];
 
     // Array that stores the value associated with each key.
-    int keyValue[MAX_KEYS];
+    short keyValue[MAX_KEYS];
 
 public:
     /**
@@ -76,7 +76,7 @@ public:
      * @param k The key (should be in the range [0, MAX_KEYS)).
      * @return Reference to the value associated with key 'k'.
      */
-    int& operator[](int k) {
+    short& operator[](int k) {
         keyVersion[k] = currentVersion;
         return keyValue[k];
     }
